requirements.txt
colormath==3.0.0
imutils==0.5.4
joblib==1.1.0
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.58
pandas==1.3.4
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
threadpoolctl==3.0.0
typing-extensions==3.10.0.2